+++ tests/underscore_component_name/_framework_box.js
@@ -1,0 +1,2 @@
+class _Framework_Box extends Jqhtml_Component {
+}